refactor: 사용자페이지 최적화
